+++ src/main/java/com/kurly/projectmaic/domain/center/dto/response/RoundResponse.java
@@ -3,7 +3,7 @@ package com.kurly.projectmaic.domain.center.dto.response;
 public record RoundResponse(
 	long roundId,
 	long centerId,
-	int centerRoundNumber,
+	long centerRoundNumber,
 	long pickTodoCount
 ) {
 }
